+++ src/main/java/com/github/arix3767/MovieRentalApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.github.arix3767;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
