+++ src/test/java/com/github/arix3767/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.github.arix3767;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
